# openapi.yaml
openapi: 3.1.0
info:
  title: Sofia Memory Plugin
  version: "1.0.0"
servers:
  - url: https://sofia-memory.onrender.com
paths:
...
  /github/repos:
    post:
      summary: List user repositories
      operationId: listRepos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Repository list
  /github/repository:
    post:
      summary: Get repository contents
      operationId: repoContents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepoRequest'
      responses:
        '200':
          description: Repository data
  /github/file:
    post:
      summary: Fetch file with lint info
      operationId: fileData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileRequest'
      responses:
        '200':
          description: File data
...
components:
  schemas:
...
    RepoRequest:
      type: object
      properties:
        token:
          type: string
        owner:
          type: string
        repo:
          type: string
        path:
          type: string
        page:
          type: integer
        per_page:
          type: integer
        fileType:
          type: string
    FileRequest:
      type: object
      properties:
        token:
          type: string
        owner:
          type: string
        repo:
          type: string
        filePath:
          type: string